Better handle legacy project system calls to set ExplicitRuleSetFilePath
